// SDKs/twilio.com/twilio_chat_v2/1.36.0/typescript/src/sdk/models/shared/chatv2servicechannelchannelwebhook.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";


export class ChatV2ServiceChannelChannelWebhook extends SpeakeasyBase {
  @Metadata({ data: "json, name=account_sid" })
  accountSid?: string;

  @Metadata({ data: "json, name=channel_sid" })
  channelSid?: string;

  @Metadata({ data: "json, name=configuration" })
  configuration?: any;

  @Metadata({ data: "json, name=date_created" })
  dateCreated?: Date;

  @Metadata({ data: "json, name=date_updated" })
  dateUpdated?: Date;

  @Metadata({ data: "json, name=service_sid" })
  serviceSid?: string;

  @Metadata({ data: "json, name=sid" })
  sid?: string;

  @Metadata({ data: "json, name=type" })
  type?: string;

  @Metadata({ data: "json, name=url" })
  url?: string;
}